Use Plugin methods for bootstrapping and setting routes.
Avoid loading separate files.

Closes #575.
