--- templates/bake/Plugin/config/routes.php.twig
@@ -1,31 +1,0 @@
-{#
-/**
- * CakePHP(tm) : Rapid Development Framework (http://cakephp.org)
- * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
- *
- * Licensed under The MIT License
- * For full copyright and license information, please see the LICENSE.txt
- * Redistributions of files must retain the above copyright notice.
- *
- * @copyright     Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
- * @link          http://cakephp.org CakePHP(tm) Project
- * @since         2.0.0
- * @license       http://www.opensource.org/licenses/mit-license.php MIT License
- */
-#}
-<?php
-declare(strict_types=1);
-
-use Cake\Routing\Route\DashedRoute;
-use Cake\Routing\RouteBuilder;
-
-/** @var \Cake\Routing\RouteBuilder $routes */
-$routes->plugin(
-    '{{ plugin }}',
-    ['path' => '/{{ routePath }}'],
-    function (RouteBuilder $builder) {
-        $builder
-            ->setRouteClass(DashedRoute::class)
-            ->fallbacks();
-    }
-);
